+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix importing a couple more bogus DAT files
 - Accept ZIP files as input for `import-dats`
 - Add a `-u` flag to `import-roms` to skip user prompts
+- Add a `-s` flag to `convert-roms` to select systems by name 
 
 # 0.18.0